listHomes: use long mtime instead of int

+++ step-core/src/main/java/com/tyndalehouse/step/core/models/HomeDirectoryInfo.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class HomeDirectoryInfo implements Serializable {
     private String path;
-    private int mtime;
+    private long mtime;
     private long size;
 
-    public HomeDirectoryInfo(String path, int mtime, long size) {
+    public HomeDirectoryInfo(String path, long mtime, long size) {
         this.path = path;
         this.mtime = mtime;
         this.size = size;
@@ -15,6 +15,6 @@ public class HomeDirectoryInfo implements Serializable {
 
     // Getters are required for JSON serialization
     public String getPath() { return path; }
-    public int getMtime() { return mtime; }
+    public long getMtime() { return mtime; }
     public long getSize() { return size; }
 }
